rename all linker's .py file
