+++ RAGchain/utils/linker/__init__.py
@@ -1,4 +1,4 @@
-from .redisdbSingleton import RedisDBSingleton
-from .dynamodbSingleton import DynamoDBSingleton
-from .localdbSingleton import LocalDBSingleton
+from .redis_linker import RedisDBSingleton
+from .dynamo_linker import DynamoDBSingleton
+from .json_linker import LocalDBSingleton
 from .base import SingletonCreationError
